add a AbstractMetadataClient and make MetadataClient to extends it
